rename `stats-fns.js` to `statsFns.ts`
