+++ src/shared/statsFns.ts
